avap_bot/requirements.txt
python-telegram-bot==21.4
gspread==6.1.0
google-auth==2.35.0
google-auth-oauthlib==1.2.1
google-auth-httplib2==0.2.0
python-dotenv==1.0.1
pytz==2024.2
difflib